+++ tests/test_cli.py
@@ -80,3 +80,7 @@ class TestCli(unittest.TestCase):
     #     self.assertEqual(mock_input.call_count, 4)
     #     self.assertEqual(mock_print.call_count, 2)
     #     self.assertEqual(mock_chess_move.call_count, 1)
+
+
+
+